翻訳の処理はtranslateモジュールに切り出した
